# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ДоВуз-Портал API
  version: 0.1.0
  description: |
    Контракт бэкенда АИС «ДоВуз-Портал». Снят со слоя `frontend/src/app/api/`
    и повторяет его как есть, включая смешанное именование полей: часть ответов
    в snake_case, часть в camelCase, query-параметры в camelCase.

    Аутентификация — HTTP-only кука сессии, которую выставляют `/auth/login`
    и `/auth/register`. Ошибки всегда приходят в формате `{"message": "..."}`.
    401 — «не вошёл» (фронт разлогинивает), 403 — «вошёл, но нельзя».
servers:
  - url: /v1
    description: Относительно хоста, с которого отдана спека; локально — http://localhost:8080/v1
security:
  - sessionCookie: []
...
paths:
...
  /directions:
    get:
      tags: [directions]
      operationId: getDirections
      summary: Все направления
      responses:
        '200':
          description: Направления
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DirectionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [directions]
      operationId: createDirection
      summary: Создать направление
      description: Только для администратора. Название уникально без учёта регистра.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
      responses:
        '201':
          description: Направление создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /directions/{id}:
    get:
      tags: [directions]
      operationId: getDirection
      summary: Направление по id
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Направление
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [directions]
      operationId: updateDirection
      summary: Переименовать направление
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
      responses:
        '200':
          description: Направление
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [directions]
      operationId: deleteDirection
      summary: Удалить направление
      description: Вместе с предметами, темами, публикациями и их файлами.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Удалено
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /news_publications:
    get:
      tags: [news]
      operationId: getNews
      summary: Лента новостей
      description: Открыта всем, новые сверху.
      security: []
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Страница новостей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationPage'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [news]
      operationId: createNews
      summary: Опубликовать новость
      description: Для модераторов и администратора.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 300
                description:
                  type: string
                  maxLength: 20000
                files:
                  type: array
                  description: До 10 файлов, каждый до 50 МБ
                  maxItems: 10
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Новость опубликована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
...
  /files/{fileNameInDirectory}:
    get:
      tags: [files]
      operationId: downloadFile
      summary: Скачать файл
      description: Вложения новостей и аватарки открыты всем, остальное — после входа.
      security:
        - {}
        - sessionCookie: []
      parameters:
        - name: fileNameInDirectory
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Содержимое файла как вложение
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /files/photos/{filename}:
    get:
      tags: [files]
      operationId: getPhoto
      summary: Картинка для `<img src>`
      description: |
        Доступ как у скачивания. JPEG, PNG, GIF и WebP отдаются для показа в браузере,
        остальные файлы — как вложение.
      security:
        - {}
        - sessionCookie: []
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Содержимое файла
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/moderators/assign:
    post:
      tags: [moderators]
      operationId: assignSubjectModerator
      summary: Назначить модератора на предмет
      description: Только для администратора. Назначить можно только пользователя с ролью «Модератор».
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/SubjectId'
      responses:
        '201':
          description: Назначение
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectModeratorResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/moderators/remove:
    delete:
      tags: [moderators]
      operationId: removeSubjectModerator
      summary: Снять модератора с предмета
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/SubjectId'
      responses:
        '204':
          description: Снят
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/moderators/subject/{id}:
    get:
      tags: [moderators]
      operationId: getSubjectModerators
      summary: Модераторы предмета
      description: Для модераторов и администратора.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Назначения на предмет
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectModeratorResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/moderators/my:
    get:
      tags: [moderators]
      operationId: getMySubjects
      summary: Мои предметы
      description: Предметы, на которые назначен текущий модератор.
      responses:
        '200':
          description: Назначения текущего модератора
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectModeratorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dovuz_session
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageNumber:
      name: pageNumber
      in: query
      description: Номер страницы, с нуля
      schema:
        type: integer
        minimum: 0
        maximum: 1000000
        default: 0
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 10
    AccountId:
      name: accountId
      in: query
      required: true
      schema:
        type: integer
    SubjectId:
      name: subjectId
      in: query
      required: true
      schema:
        type: integer
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Нужно войти в систему
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PayloadTooLarge:
      description: Файл слишком большой
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Name:
      type: string
      minLength: 1
      maxLength: 300
    PageMeta:
      type: object
      required: [total_pages, total_size, page_number, page_size]
      properties:
        total_pages:
          type: integer
        total_size:
          type: integer
        page_number:
          type: integer
        page_size:
          type: integer
...
    NameRequest:
      type: object
      required: [name]
      properties:
        name:
          $ref: '#/components/schemas/Name'
    DirectionResponse:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
...
    FileInfo:
      type: object
      required: [id, file_name_in_directory, initial_file_name]
      properties:
        id:
          type: integer
        file_name_in_directory:
          type: string
          description: Имя в хранилище, по нему файл скачивается
        initial_file_name:
          type: string
          description: Имя, под которым файл загрузили
    PublicationResponse:
      type: object
      required: [id, title, description, created_at, nickname, files]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        nickname:
          type: [string, 'null']
          description: Автор; `null`, если его учётную запись удалили
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileInfo'
...
    PublicationPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/PublicationResponse'
...
    SubjectModeratorResponse:
      type: object
      required: [id, accountId, nickname, subjectId, subjectName]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        nickname:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string